Moved Footer and Feedback Button to their own component files
