+++ client/src/App.css
@@ -282,6 +282,8 @@ That is 0.71 x 1 */
   cursor: pointer;
 }
 
+/* FOOTER */
+
 .footer {
   position: absolute;
   right: 0;
